refactor(scm): moved scm errors to new scm_errors.go file
